Initial async NM client patch
